requirements.txt
google-api-python-client
google-auth-oauthlib
pandas
numpy
textblob
nltk
emoji
matplotlib
seaborn
